backend/requirements.txt
dnspython==2.8.0
motor==3.6.0
pymongo>=4.9,<4.10
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
openai==1.82.0
apscheduler==3.10.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
tenacity==9.0.0
structlog==24.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
